tested with device configuration 3.2
